+++ web/packages/new/photos/components/CollectionSelector.tsx
@@ -38,7 +38,6 @@ import {
 import { includes } from "ente-utils/type-guards";
 import { t } from "i18next";
 import React, { useEffect, useMemo, useRef, useState } from "react";
-import { useSettingsSnapshot } from "./utils/use-snapshot";
 
 export type CollectionSelectorAction =
     | "upload"
@@ -141,8 +140,6 @@ export const CollectionSelector: React.FC<CollectionSelectorProps> = ({
 }) => {
     // Make the dialog fullscreen if the screen is <= the dialog's max width.
     const isFullScreen = useMediaQuery("(max-width: 490px)");
-    const { isInternalUser } = useSettingsSnapshot();
-    const canUseSharedAlbumUpload = isInternalUser;
 
     const [searchTerm, setSearchTerm] = useState("");
     const [sortBy, setSortBy] =
@@ -169,11 +166,9 @@ export const CollectionSelector: React.FC<CollectionSelectorProps> = ({
                 } else if (attributes.action == "add") {
                     return canAddToCollection(cs) && cs.type != "userFavorites";
                 } else if (attributes.action == "upload") {
-                    const canUploadToCollection = canUseSharedAlbumUpload
-                        ? canAddToCollection(cs)
-                        : canMoveToCollection(cs);
                     return (
-                        (canUploadToCollection || cs.type == "uncategorized") &&
+                        (canAddToCollection(cs) ||
+                            cs.type == "uncategorized") &&
                         cs.type != "userFavorites"
                     );
                 } else if (attributes.action == "restore") {
@@ -206,14 +201,7 @@ export const CollectionSelector: React.FC<CollectionSelectorProps> = ({
         }
 
         setFilteredCollections(collections);
-    }, [
-        collectionSummaries,
-        attributes,
-        open,
-        onClose,
-        sortBy,
-        canUseSharedAlbumUpload,
-    ]);
+    }, [collectionSummaries, attributes, open, onClose, sortBy]);
 
     const searchFilteredCollections = useMemo(() => {
         if (!searchTerm.trim()) {